+++ sync-for-payables/src/sdk/models/operations/uploadbillattachment.ts
@@ -21,19 +21,28 @@ export class UploadBillAttachmentRequest extends SpeakeasyBase {
     requestBody?: UploadBillAttachmentRequestBody;
 
     /**
-     * Unique identifier for a bill
+     * Unique identifier for a bill.
      */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=billId" })
     billId: string;
 
+    /**
+     * Unique identifier for a company.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=companyId" })
     companyId: string;
 
+    /**
+     * Unique identifier for a connection.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=connectionId" })
     connectionId: string;
 }
 
 export class UploadBillAttachmentResponse extends SpeakeasyBase {
+    /**
+     * HTTP response content type for this operation
+     */
     @SpeakeasyMetadata()
     contentType: string;
 
@@ -43,9 +52,15 @@ export class UploadBillAttachmentResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;
 
+    /**
+     * HTTP response status code for this operation
+     */
     @SpeakeasyMetadata()
     statusCode: number;
 
+    /**
+     * Raw HTTP response; suitable for custom response parsing
+     */
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
 }